+++ src/cachelib/uwsgi.py
@@ -1,9 +1,5 @@
+import pickle
 import platform
-
-try:
-    import cPickle as pickle
-except ImportError:  # pragma: no cover
-    import pickle
 
 from cachelib.base import BaseCache